backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
geopy==2.4.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
